+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 # Version 0.0.3 (wip)
+- Planets not building warning includes link to build page
 - Fixed radar page system shortcut logic
 - Research page cosmetics
 - Popup cosmetics