+++ CHANGELOG.md
@@ -1,17 +1,18 @@
 
-# Version 0.0.3 (wip)
+# Version 0.0.3
+- Updated fleet scores formula (see fleet page)
 - Planets not building warning includes link to build page
 - Fixed radar page system shortcut logic
 - Research summary + copy to cliboard
 - Popup cosmetics
 - Bugfix fleet scan with player without alliance
 
-# Version 0.0.2 (31st Aug'23)
+# Version 0.0.2
 - Fixed popup links
 - Fleet list quick filter by sector
 - Bugfix fleet list quick filter
 - Copy to cliboard in fleet page
 - Bugfix surface scan
 
-# Version 0.0.1 (29th Aug'23)
+# Version 0.0.1
 - initial version